dclde_2026/requirements.txt
torch
torchvision
torchaudio
ultralytics
transformers
librosa
pandas
gcsfs
google-cloud-storage
scikit-learn
pyyaml
torchmetrics
tqdm
scikit-image
soundfile
opencv-python
scipy
audiomentations
wandb

